+++ core/processor.py
@@ -1,7 +1,0 @@
-processor.start()
-
-processor.stop()
-
-processor.pause()
-
-processor.resume()