+++ main.py
@@ -1,0 +1,9 @@
+from ui.main_window import MainWindow
+from PySide6.QtWidgets import QApplication
+
+
+if __name__ == "__main__":
+    app = QApplication([])
+    window = MainWindow()
+    window.show()
+    app.exec()